New green skybox for test map. Added in func_no_build brush ent. Added in obj_resourcepump and obj_sandbag_bunker. Added simple little resource_chunk model. Fixed attachments on human_obj_powerpack. Updated weapon_combat_burstrifle to use new firing sound. Updated technologytree with base tech sets and added default weapon loadout. Added minimap (minor bug with coords after spawning, just zoom it to workaround.) Updated ClientSchemeAlien to mirror ClientSchemeHuman. Updated test level with some minor geometric changes. Removed v_slam model, this is included with HL2DM duh. Fixed blueflare2 material. Fixed an issue with footprint materials.

+++ mp/src/game/client/fortress/hud_damageindicator.cpp
@@ -158,7 +158,7 @@ void CHudDamageIndicator::GetDamagePosition( const Vector &vecDelta, float flRad
 	*ypos = flRadius * -front;
 
 	// Get the rotation (yaw)
-	*flRotation = (float)(atan2((double)*xpos,(double)*ypos)+M_PI);
+	*flRotation = atan2((float)*xpos,(float)*ypos)+M_PI;
 	*flRotation *= 180 / M_PI;
 
 	float yawRadians = -(*flRotation) * M_PI / 180.0f;
